# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WhatPay API",
    "description": "Plataforma de pagos peer-to-peer integrada con WhatsApp para Chile.\n\n## Autenticación\nLa mayoría de endpoints requieren un JWT Bearer token obtenido en `POST /api/v1/users/login`.\nLos endpoints admin requieren el header `x-admin-key`.\n\n## Flujo típico\n1. Registro vía bot WhatsApp (conversacional) o `POST /api/v1/users/register`\n2. `POST /api/v1/users/login` → PIN → JWT (30 min)\n3. `POST /api/v1/topup/webpay` → recarga wallet con Transbank\n4. `POST /api/v1/payments/pay` → envía pago P2P\n5. `POST /api/v1/payments/links` → crea link de cobro para comercios\n\n## Niveles KYC\n| Nivel | Límite por tx | Límite mensual |\n|-------|--------------|----------------|\n| `BASIC` | $50.000 | $200.000 |\n| `INTERMEDIATE` | $500.000 | $2.000.000 |\n| `FULL` | $2.000.000 | Sin límite |\n\n## Seguridad\n- PIN almacenado como bcrypt hash (cost 12)\n- RUT almacenado como HMAC-SHA256 (nunca en claro)\n- Bloqueo tras 3 intentos fallidos de PIN (15 min)\n- Rate limiting por IP en todos los endpoints\n- Todas las transferencias usan SELECT FOR UPDATE (sin double-spending)\n",
    "version": "0.1.0",
    "contact": {
      "name": "Pablo Cussen",
      "url": "https://cussen.cl",
      "email": "pablo@cussen.cl"
    },
    "license": {
      "name": "Portfolio Project",
      "url": "https://cussen.cl/whatpay"
    }
  },
  "servers": [
    {
...
      "description": "Producción (GCP Cloud Run — Santiago)"
...
    {
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido en POST /api/v1/users/login"
      },
      "adminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key",
        "description": "API key de administrador"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Datos inválidos."
          }
        }
      },
...
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "waId": {
            "type": "string",
            "example": "+56912345678"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "kycLevel": {
            "type": "string",
            "enum": [
              "BASIC",
              "INTERMEDIATE",
              "FULL"
            ]
          },
          "biometricEnabled": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "balance": {
            "type": "object",
            "properties": {
              "amount": {
                "type": "integer",
                "example": 25000
              },
              "formatted": {
                "type": "string",
                "example": "$25.000 CLP"
              }
            }
          },
          "stats": {
            "type": "object",
            "properties": {
              "totalSent": {
                "type": "integer"
              },
              "totalReceived": {
                "type": "integer"
              },
              "txCount": {
                "type": "integer"
              }
            }
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/users/register": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Registrar usuario",
        "security": [],
        "description": "Crea una cuenta nueva. RUT validado y almacenado como HMAC-SHA256. PIN como bcrypt (cost 12). Rate limit: 3/hora/IP.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "waId",
                  "rut",
                  "pin"
                ],
                "properties": {
                  "waId": {
                    "type": "string",
                    "example": "+56912345678"
                  },
                  "rut": {
                    "type": "string",
                    "example": "12345678-9"
                  },
                  "pin": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 6,
                    "example": "483920"
                  },
                  "name": {
                    "type": "string",
                    "example": "Juan Pérez"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado",
            "content": {
              "application/json": {
                "example": {
                  "token": "eyJ...",
                  "user": {
                    "id": "uuid",
                    "kycLevel": "BASIC"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Autenticar con PIN",
        "security": [],
        "description": "Verifica PIN con bcrypt → JWT (30 min). Bloqueo tras 3 intentos fallidos (15 min).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "waId",
                  "pin"
                ],
                "properties": {
                  "waId": {
                    "type": "string",
                    "example": "+56912345678"
                  },
                  "pin": {
                    "type": "string",
                    "example": "483920"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT generado",
            "content": {
              "application/json": {
                "example": {
                  "token": "eyJ...",
                  "user": {
                    "id": "uuid",
                    "kycLevel": "BASIC"
                  }
                }
              }
            }
          },
          "401": {
            "description": "PIN incorrecto"
          },
          "423": {
            "description": "Cuenta bloqueada (15 min)"
          }
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Perfil del usuario autenticado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Perfil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
...
  }
}